Moves factory functions and exceptions into dedicated modules
