+++ sndata/csp/dr1/_data_download.py
@@ -4,9 +4,10 @@
 """This module defines functions for downloading data."""
 
 from . import _meta as meta
+from ... import _factory_funcs as factory
 from ... import _utils as utils
 
-delete_module_data = utils.factory_delete_module_data(meta.data_dir)
+delete_module_data = factory.factory_delete_module_data(meta.data_dir)
 
 
 def download_module_data(force=False):
@@ -19,7 +20,6 @@ def download_module_data(force=False):
     # Download data tables
     if (force or not meta.table_dir.exists()) \
             and utils.check_url(meta.table_url):
-
         print('Downloading data tables...')
         utils.download_tar(
             url=meta.table_url,
@@ -29,7 +29,6 @@ def download_module_data(force=False):
     # Download spectra
     if (force or not meta.spectra_dir.exists()) \
             and utils.check_url(meta.spectra_url):
-
         print('Downloading spectra...')
         utils.download_tar(
             url=meta.spectra_url,
