+++ sndata/exceptions.py
@@ -1,0 +1,25 @@
+#!/usr/bin/env python3.7
+# -*- coding: UTF-8 -*-
+
+"""This module defines custom exceptions."""
+
+
+class NoDownloadedData(Exception):
+    def __init__(self, *args, **kwargs):
+        default_message = \
+            'Data has not been downloaded for this module / data release.'
+
+        if not (args or kwargs):
+            args = (default_message,)
+
+        super().__init__(*args, **kwargs)
+
+
+class InvalidObjId(Exception):
+    def __init__(self, *args, **kwargs):
+        default_message = 'The provided object Id is not valid.'
+
+        if not (args or kwargs):
+            args = (default_message,)
+
+        super().__init__(*args, **kwargs)
